Fixed Custom location ui #142

+++ app/src/main/java/com/recep/hunt/model/usersList/Data.kt
@@ -38,6 +38,6 @@ data class Data(
     val socialInfo : SocialInfo,
     val user_profile_image : ArrayList<ProfileImage>,
     val totalMatching : Float,
-    val totalMeeting : Int,
+    val totalMeeting : Float,
     val user_swipe : String
 )
